Refactor client source layout
Slightly more sophisticated client side layout. I suspect this will
get refactored again in the not too distant future but this seems like a
good starting point.
