+++ src/client/components/config/config.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { FormGroup, FormControl, ControlLabel } from 'react-bootstrap';
 
-import Modal from './modal/modal';
+import Modal from '../modal/modal';
 import { modal } from './config.css.js'
 
 const ConfigFolder = ({name, directory}) =>
