+++ src/client/index.js
@@ -1,8 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './app';
+import Layout from './layout/layout';
+
+import './index.css.js';
 
 ReactDOM.render(
-  <App />,
+  <Layout />,
   document.getElementById('root')
 );
